:sparkles: Swagger 추가

+++ src/main/java/kr/co/selab/selabtodo/service/TodoService.java
@@ -45,6 +45,9 @@ public class TodoService {
 
     @Transactional
     public void deleteTodo(Long id) {
+        if(!todoRepository.existsById(id)) {
+            throw new NotExistsTodoException();
+        }
         todoRepository.deleteById(id);
     }
